fix: rename/relocate hooks and feature flags
