+++ apps/web/src/features/targetedFeatures/constants.ts
@@ -1,0 +1,6 @@
+import { FEATURES } from '@/utils/chains'
+
+export const TARGETED_FEATURES = [
+  { id: 3, feature: FEATURES.TARGETED_NESTED_SAFES },
+  { id: 4, feature: FEATURES.TARGETED_MASS_PAYOUTS },
+] as const
