+++ apps/web/src/utils/chains.ts
@@ -41,8 +41,8 @@ export enum FEATURES {
   BRIDGE = 'BRIDGE',
   RENEW_NOTIFICATIONS_TOKEN = 'RENEW_NOTIFICATIONS_TOKEN',
   TX_NOTES = 'TX_NOTES',
-  NESTED_SAFES = 'NESTED_SAFES',
-  MASS_PAYOUTS = 'MASS_PAYOUTS',
+  TARGETED_NESTED_SAFES = 'TARGETED_NESTED_SAFES',
+  TARGETED_MASS_PAYOUTS = 'TARGETED_MASS_PAYOUTS',
 }
 
 export const FeatureRoutes = {
